+++ src/index.ts
@@ -69,8 +69,11 @@ export {
     buildDummyTransferInput,
     createNoteDisclosureKey,
     decodeNoteDisclosureKey,
+    CircuitVersionResolver,
 } from './shielded-pool/index';
 export type { NoteDisclosure } from './shielded-pool/protocol/NoteDisclosure';
+export type { ResolvedSpendVersion } from './shielded-pool/index';
+export { CURRENT_CIRCUIT_VERSION } from './shielded-pool/protocol/types';
 export { BN254_R, BABYJUB_SUBORDER } from './utils/crypto-constants';
 export { randomBlinding } from './utils/blinding';